+++ budgets/models.py
@@ -46,8 +46,8 @@ class Budget(models.Model):
 class BudgetAssignment(models.Model):
     budget = models.ForeignKey(Budget, on_delete=models.CASCADE)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
-    budget_amount = models.DecimalField(max_digits=10, decimal_places=2)
-    expense_amount = models.DecimalField(max_digits=10, decimal_places=2)
+    budget_amount = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    expense_amount = models.DecimalField(max_digits=10, decimal_places=2, default=0)
 
     class Meta:
         verbose_name = "Asignacion de presupuesto"