+++ expenses/api/accounts.py
@@ -1,10 +1,16 @@
-from django.conf import settings
-from django.db.models import Q
 from django.http import JsonResponse
 
+from rest_framework import viewsets
 from rest_framework.views import APIView
 
+from expenses.models import Account
 from expenses.utils import change_account_from_assoc
+from expenses.serializers import AccountSerializer
+
+
+class AccountViewSet(viewsets.ModelViewSet):
+    queryset = Account.objects.all()
+    serializer_class = AccountSerializer
 
 
 class AccountAssociationView(APIView):