+++ expenses_site/urls.py
@@ -18,8 +18,8 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("", include("expenses.urls")),
+    path("expenses/", include("expenses.urls")),
+    path("budgets/", include("budgets.urls")),
     path("api/", include("expenses.api.urls")),
-    path("budget/", include("budgets.urls")),
     path("__debug__/", include("debug_toolbar.urls")),
 ]